Server/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-jwt-extended
psycopg2-binary
marshmallow
marshmallow-sqlalchemy
gunicorn
